an other refactoring!
