+++ scripts/hammerstone/ddapi/ddapiUtils.lua
@@ -1,8 +1,8 @@
 --- objectUtils.lua
---- Utility methods, generally used by the objectManager
+--- Utility methods, generally used by the ddapiManager
 --- @author earmuffs, SirLich
 
-local objectUtils = {}
+local ddapiUtils = {}
 
 -- Hammestone
 local log = mjrequire "hammerstone/logging"
@@ -11,7 +11,7 @@ local utils = mjrequire "hammerstone/utils/utils"
 local runOnceGuards = {}
 --- Guards against the same code being run multiple times.
 -- @param id string - The unique identifier for this guard.
-function objectUtils:runOnceGuard(id)
+function ddapiUtils:runOnceGuard(id)
 	if runOnceGuards[id] == nil then
 		runOnceGuards[id] = true
 		return false
@@ -22,7 +22,7 @@ end
 --- Returns the type, or nil if not found. Logs error.
 -- @param tbl The table where the key can be found in. e.g., gameObject.types
 -- @param key The key such as "inca:rat_skull" which will be cast to type.
-function objectUtils:getType(tbl, key, displayAlias)
+function ddapiUtils:getType(tbl, key, displayAlias)
 	if displayAlias == nil then
 		displayAlias = tostring(key)
 	end
@@ -30,14 +30,14 @@ function objectUtils:getType(tbl, key, displayAlias)
 	if tbl[key] ~= nil then
 			return tbl[key]   
 	end
-	return objectUtils:logMissing(displayAlias, key, tbl)
+	return ddapiUtils:logMissing(displayAlias, key, tbl)
 end
 
 
 --- Returns the index of a type, or nil if not found.
 -- @param tbl The table where the index can be found in. e.g., gameObject.types
 -- @param key The key such as "inca:rat_skull" where which will be cast to type.
-function objectUtils:getTypeIndex(tbl, key, displayAlias)
+function ddapiUtils:getTypeIndex(tbl, key, displayAlias)
 	if displayAlias == nil then
 		displayAlias = tostring(key)
 	end
@@ -45,11 +45,11 @@ function objectUtils:getTypeIndex(tbl, key, displayAlias)
 	if tbl[key] ~= nil then
 		return tbl[key].index
 	end
-	return objectUtils:logMissing(displayAlias, key, tbl)
+	return ddapiUtils:logMissing(displayAlias, key, tbl)
 end
 
 local logMissingTables = {}
-function objectUtils:logMissing(displayAlias, key, tbl)
+function ddapiUtils:logMissing(displayAlias, key, tbl)
 	if logMissingTables[tbl] == nil then
 
 		table.insert(logMissingTables, tbl)
@@ -73,11 +73,11 @@ function objectUtils:logMissing(displayAlias, key, tbl)
 	end
 end
 
-function objectUtils:logNotImplemented(featureName)
+function ddapiUtils:logNotImplemented(featureName)
 	log:schema("ddapi", "    WARNING: " .. featureName .. " is used but it is yet to be implemented")
 end
 
-function objectUtils:debug(identifier, config, tbl)
+function ddapiUtils:debug(identifier, config, tbl)
 	if config.debug then
 		log:schema("ddapi", "DEBUGGING: " .. identifier)
 		log:schema("ddapi", "Config:")
@@ -87,4 +87,39 @@ function objectUtils:debug(identifier, config, tbl)
 	end
 end
 
-return objectUtils
+----------------------------
+-- utils for locale
+----------------------------
+function ddapiUtils:getBuildIdentifier(identifier)
+	return "build_" .. identifier
+end
+
+function ddapiUtils:getNameKey(prefix, identifier)
+	return prefix .. "_" .. identifier
+end
+
+function ddapiUtils:getPluralKey(prefix, identifier)
+	return prefix .. "_" .. identifier .. "_plural"
+end
+
+function ddapiUtils:getNameLocKey(identifier)
+	return "object_" .. identifier
+end
+
+function ddapiUtils:getPluralLocKey(identifier)
+	return "object_" .. identifier .. "_plural"
+end
+
+function ddapiUtils:getSummaryLocKey(identifier)
+	return "object_" .. identifier .. "_summary"
+end
+
+function ddapiUtils:getInProgressKey(prefix, identifier)
+	return prefix .. "_" .. identifier .. "_inProgress"
+end
+
+function ddapiUtils:getDescriptionKey(prefix, identifier)
+	return prefix .. "_" ..identifier .. "_description"
+end
+
+return ddapiUtils
