+++ scripts/hammerstone/ddapi/entityManagers/builderManager.lua
@@ -1,0 +1,12 @@
+local builderManager = {
+    settings = {
+        isGlobal = true,
+			configPath = "/hammerstone/builders/",
+    }, 
+    loaders = {}
+}
+
+function builderManager:init()
+end
+
+return builderManager
